Listen for Windows events when focusing/unfocusing browser panels
Follow-up to https://github.com/WizardCM/obs-browser/tree/fix-focus

Since finding out that CEF doesn't provide any kind of focus/unfocus
events itself, I've spent an ungodly amount of time trying to come up
with alternatives. In the end, I now have a function that is triggered
when *Windows* detects that the CEF browser is focused/unfocused.

The current flaw with this design, is that a listener is added for every
panel, but CEF only has one parent window. Therefore, the message is
duplicated by the number of browser docks you have open. Not ideal.

Next steps:
  1. Move HookCallback into somewhere more useful - specifically into somewhere
     where I can trigger Qt events/functions from it. Depending on the solution
     to the message duplication, this'll either be somewhere in QCefWidgetInternal,
     or somewhere else entirely. Haven't figured that out yet.
  2. Within HookCallback, I need to force-update Qt from Qt::ApplicationActive.
     Ideally, I should be able to just unfocus any old Qt object, or the Qt window,
     or something even higher up. I really, really hope it'll be that easy.
     But I have no idea if it will.
  3. Combine this with the fix-focus branch. This will mean that some code here or
     some code there will need to be removed, as there is a bit of duplication
     between the two, as both are different methods of trying to achieve the same goal.